requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
pymongo==4.6.3
apscheduler==3.10.4
python-dotenv==1.0.0
pytz==2023.3
pydantic==2.4.2
typing-extensions==4.8.0
gunicorn==21.2.0